nnet3: Refactoring nnet-optimize.{h,cc} slightly by adding new header/.cc file; other code refactoring.
