+++ src/nnet3/Makefile
@@ -22,13 +22,14 @@ OBJFILES = nnet-common.o nnet-compile.o nnet-component-itf.o \
   nnet-example.o nnet-nnet.o nnet-compile-utils.o \
   nnet-utils.o nnet-compute.o nnet-test-utils.o nnet-analyze.o \
   nnet-example-utils.o nnet-training.o \
-  nnet-diagnostics.o nnet-combine.o nnet-am-decodable-simple.o
+  nnet-diagnostics.o nnet-combine.o nnet-am-decodable-simple.o \
+  nnet-optimize-utils.o
 
 LIBNAME = kaldi-nnet3
 
 ADDLIBS = ../thread/kaldi-thread.a ../lat/kaldi-lat.a ../gmm/kaldi-gmm.a \
       ../hmm/kaldi-hmm.a ../tree/kaldi-tree.a ../transform/kaldi-transform.a \
       ../cudamatrix/kaldi-cudamatrix.a ../matrix/kaldi-matrix.a \
-      ../base/kaldi-base.a  ../util/kaldi-util.a 
+      ../base/kaldi-base.a  ../util/kaldi-util.a
 
 include ../makefiles/default_rules.mk
